% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\name{buildModel,TrenaValidator-method}
\alias{buildModel,TrenaValidator-method}
\alias{buildModel}
\title{return a trena model}
\usage{
\S4method{buildModel}{TrenaValidator}(obj, geneCutoff = 0)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{geneCutoff}{numeric, betweeen 0 and 1, controls inclusion of weaker TFs in model}
}
\value{
a data.frame
}
\description{
return a trena model
}
