% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{fallbackEnhancers,TrenaValidator-method}
\alias{fallbackEnhancers,TrenaValidator-method}
\alias{fallbackEnhancers}
\title{a simple single 10k regulatory region used if no geneHancers values are avaialable}
\usage{
\S4method{fallbackEnhancers}{TrenaValidator}(obj)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}
}
\value{
data.frame
}
\description{
a simple single 10k regulatory region used if no geneHancers values are avaialable
}
