% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\name{findEnhancers,TrenaValidator-method}
\alias{findEnhancers,TrenaValidator-method}
\alias{findEnhancers}
\title{query and store the GeneHancer regions for the targetGene}
\usage{
\S4method{findEnhancers}{TrenaValidator}(obj, eliteOnly)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{eliteOnly}{keep all regions, on only those with two attestations}
}
\value{
nothing
}
\description{
query and store the GeneHancer regions for the targetGene
}
