% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{getEnhancerTable,TrenaValidator-method}
\alias{getEnhancerTable,TrenaValidator-method}
\alias{getEnhancerTable}
\title{return the enhancers}
\usage{
\S4method{getEnhancerTable}{TrenaValidator}(obj)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}
}
\value{
a data.frame
}
\description{
return the enhancers
}
