% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{getSimplePromoter,TrenaValidator-method}
\alias{getSimplePromoter,TrenaValidator-method}
\alias{getSimplePromoter}
\title{a simple single regulatory region oritened around the tss}
\usage{
\S4method{getSimplePromoter}{TrenaValidator}(obj, upstream = 5000,
  downstream = 5000)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{upstream}{numeric, default 5kb}

\item{downstream}{numeric, default 5kb}
}
\value{
data.frame
}
\description{
a simple single regulatory region oritened around the tss
}
