% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{getTFBS.bioc,TrenaValidator-method}
\alias{getTFBS.bioc,TrenaValidator-method}
\alias{getTFBS.bioc}
\title{return the TFBS using the bioc pwm match, trena MotifMatcher class}
\usage{
\S4method{getTFBS.bioc}{TrenaValidator}(obj, tbl.regions, match.threshold,
  conservation.threshold, pwms)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{tbl.regions}{a data.frame with chrom, start, end columns}

\item{match.threshold}{numeric, between 0 and 100}

\item{conservation.threshold}{numeric, between zero and one}

\item{pwms}{a list of motifs from MotifDb}
}
\value{
a data.frame
}
\description{
return the TFBS using the bioc pwm match, trena MotifMatcher class
}
