% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{getTFBS,TrenaValidator-method}
\alias{getTFBS,TrenaValidator-method}
\alias{getTFBS}
\title{return the TFBS}
\usage{
\S4method{getTFBS}{TrenaValidator}(obj, tbl.regions, fimo.threshold,
  conservation.threshold, pwmFile)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}
}
\value{
a data.frame
}
\description{
return the TFBS
}
