% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{setMatrix,TrenaValidator-method}
\alias{setMatrix,TrenaValidator-method}
\alias{setMatrix}
\title{set the matrix to use in model building}
\usage{
\S4method{setMatrix}{TrenaValidator}(obj, matrix)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{matrix}{a properly normalized numeric matrix with geneSymbol rownames}
}
\value{
noting
}
\description{
set the matrix to use in model building
}
