% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{setRegulatoryRegionsTable,TrenaValidator-method}
\alias{setRegulatoryRegionsTable,TrenaValidator-method}
\alias{setRegulatoryRegionsTable}
\title{set the regulatory regions table}
\usage{
\S4method{setRegulatoryRegionsTable}{TrenaValidator}(obj, tbl.reg)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{tbl.reg}{A data.frame}
}
\value{
noting
}
\description{
set the regulatory regions table
}
