% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{class}
\name{TrenaValidator-class}
\alias{TrenaValidator-class}
\alias{.TrenaValidator}
\alias{TrenaValidator}
\title{TrenaValidator-class}
\usage{
TrenaValidator(TF, targetGene, mtx, tbl.benchmark, quiet = TRUE)
}
\value{
An object of the TrenaValidator class
}
\description{
A TF/targetGene/tissue combination
}
