% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\name{buildModel,TrenaValidator-method}
\alias{buildModel,TrenaValidator-method}
\alias{buildModel}
\title{return a trena model}
\usage{
\S4method{buildModel}{TrenaValidator}(obj)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}
}
\value{
a data.frame
}
\description{
return a trena model
}
