% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaValidator.R
\docType{methods}
\name{getTFBS,TrenaValidator-method}
\alias{getTFBS,TrenaValidator-method}
\alias{getTFBS}
\title{return the TFBS}
\usage{
\S4method{getTFBS}{TrenaValidator}(obj, tbl.regions, fimo.threshold,
  conservation.threshold, pwmFile)
}
\arguments{
\item{obj}{An instance of the TrenaValidator class}

\item{tbl.regions}{a data.frame with chrom, start, end columns}

\item{fimo.threshold}{numeric, 1e-4 for example}

\item{conservation.threshold}{numeric, between zero and one}

\item{pwmFile}{character string, a full or relative path to a file in meme format}
}
\value{
a data.frame
}
\description{
return the TFBS
}
